requirements.txt
fire==0.1.2
libtmux==0.7.7
PyYAML==3.12
termcolor==1.1.0